refactor: split ansi.ts god file into 8 focused modules
Break down the 3560-line ansi.ts monolith into 8 focused, maintainable
modules organized by feature domain. Maintains 100% backward compatibility
through re-exports.

New module structure:
- ansi/constants.ts (27 lines) - Core ANSI escape constants
- ansi/colors.ts (194 lines) - Color types and SGR codes
- ansi/cursor.ts (373 lines) - Cursor movement and styling
- ansi/text.ts (451 lines) - Text styling and screen control
- ansi/mouse.ts (538 lines) - Mouse protocols and locator
- ansi/features.ts (456 lines) - Modern features (sync, paste, clipboard, tmux)
- ansi/charset.ts (575 lines) - Character sets and box drawing
- ansi/advanced.ts (991 lines) - Advanced/niche features (window, hyperlink, media, rectangle)
- ansi/index.ts (67 lines) - Central re-export hub

Benefits:
- Improved maintainability: Each module has single responsibility
- Better discoverability: Related functions grouped logically
- Faster loading: Smaller modules enable better tree-shaking
- Easier testing: Feature areas can be tested in isolation
- Preserved JSDoc: All documentation and examples maintained

The original ansi.ts now re-exports from ./ansi/index with deprecation
notice for future removal.

Issue: #1077
