+++ src/terminal/ansi/index.ts
@@ -5,6 +5,41 @@
  * @internal
  */
 
+// Module 8: Advanced
+export type { HyperlinkOptions, HyperlinkProtocol, MediaCopyModeValue } from './advanced';
+export {
+	HYPERLINK_ALLOWED_PROTOCOLS,
+	hyperlink,
+	isHyperlinkAllowed,
+	MediaCopyMode,
+	mediaCopy,
+	rectangle,
+	windowOps,
+} from './advanced';
+// Module 7: Charset
+export type { BoxDrawingSet, CharacterSetId, CharacterSetRegister } from './charset';
+export { boxDrawing, charset, DEC_SPECIAL_GRAPHICS, UNICODE_TO_ASCII } from './charset';
+// Module 2: Colors
+export type { BasicColor, Color, Color256, RGBColor } from './colors';
+export { basicBgColors, basicFgColors, isBasicColor, isRGBColor, SGR, sgr } from './colors';
+// Module 1: Constants
+export { BEL, CSI, DCS, ESC, OSC, ST } from './constants';
+// Module 3: Cursor
+export type { CursorShapeType } from './cursor';
+export { CursorShape, cursor } from './cursor';
+// Module 6: Features
+export type { ClipboardSelectionType } from './features';
+export {
+	bracketedPaste,
+	ClipboardSelection,
+	clipboard,
+	DEFAULT_CLIPBOARD_MAX_SIZE,
+	sync,
+	tmux,
+} from './features';
+// Module 5: Mouse
+export type { LocatorButtonValue, LocatorEventValue } from './mouse';
+export { LocatorButton, LocatorEvent, locator, MouseMode, mouse } from './mouse';
 // Parser exports
 export type { Attribute, CodeAttrOptions, InternalColor, OutputColorDepth } from './parser';
 export {
@@ -28,3 +63,5 @@ export {
 	unpackRgb,
 	visibleLength,
 } from './parser';
+// Module 4: Text and Screen
+export { screen, style, title } from './text';
